+++ build.gradle
@@ -10,7 +10,7 @@ subprojects {
     apply plugin: 'maven'
 
     group = "com.github.otah"
-    version = "0.1-SNAPSHOT"
+    version = "0.2-observables-SNAPSHOT"
 
     sourceCompatibility = 1.8
