+++ hap-monix/src/main/scala/com/github/otah/hap/monix/ObservableCharacteristic.scala
@@ -1,4 +1,4 @@
-package com.github.otah.hap.observable
+package com.github.otah.hap.monix
 
 import com.github.otah.hap.api.Characteristic
 import monix.eval.Task
