+++ hap-reactivex/src/main/scala/com/github/otah/hap/reactivex/ObservableCharacteristic.scala
@@ -1,4 +1,4 @@
-package com.github.otah.hap.observable
+package com.github.otah.hap.reactivex
 
 import com.github.otah.hap.api.Characteristic
 import rx.lang.scala.Observable
